fix(server): prevent protocol error by removing null reason
If the opponent violated, the Game used to crash which also crashed the
TestClient, leading to confusing results and inconsistent tests.

This was because the Game tried to instantiate a PlayerScore with a null
reason, which is disallowed and thus produced an exception. So the reason
now defaults to an empty string instead.