refactor(sdk): convert RoundBasedGameInstance to Kotlin
This is a crucial change; as super class for Games it provides several
methods handling MoveRequests and as such has to work properly.
The Java/Kotlin mix is a potential source of error whereever on calls
the other.
Since this class, however, is so deep-seated, converting it requires
some changes and conversions in other classes as well.

Also, IMPORTANTLY, two server tests broke - the gui still runs fine with
these changes, but I can't guarantee (and can't fix the tests either).