fix(server): do not terminate connection on error
Previously, the server would terminate the connection to a client upon error,
now it only notifies the client.

Suppose the following scenario:
An administrative Client sends a CancelRequest,
but the targeted room is already closed while the message is underway.
The administrative client would have previously been kicked by the server
for no good reason, as a missing room constitutes an error.