feat: remove obsolete game logic and introduce new game logic
Removed the old "Tile" data class which contained information that was not necessary anymore. The new logic will handle the gameboard in "segments" in such a way that it considers the direction of the field.

Additionally, I introduced the generation of segments and added this functionality to the board. This is necessary for the new way the game should be handled in which the board is split into different segments.

Changed "FieldType" from a sealed class to an enum which simplifies the usage and makes the use of when expressions more clear.

In addition, the types were adapted and added according to new requirements.

Furthermore, added a function for creating a random HexDirection in HexDirection for convenience reasons.

This is a major change in the underlying game logic which is necessary for future development and features.