refactor: change Segment and Board properties to mutable values
Changed both the 'nextSegment' property in the Segment class and the 'segments' property in the Board class from read-only to mutable. This modification was done to accommodate changes in the program logic, mainly the implementation of new features like serialization and making it easier to determine the start point of segments. Made changes in the 'initSegment' function to reflect the new mutable properties as well.