feat: added turnCountTo method and isPassable check
This commit introduces changes in two source files, namely: Coordinates.kt and Field.kt. In the Coordinates.kt file, a method named turnCountTo() has been added to the HexDirection enum and the opposite() function was slightly optimized to use the values() method instead of HexDirection.values(). The turnCountTo() function calculates the minimum number of turns to redirect current HexDirection towards the target HexDirection.

In the Field.kt file a new method called isPassable() was added to the Field class, which determines whether a field of a specific type can be passed or not based on its type. This can be useful in different situations like movement validation, game logic, etc.