refactor: refactor Ship and Board classes, and implement IPiece interface.
The Ship class was refactored to implement the newly introduced IPiece interface. The interface adds the properties of id, position, and team to the Ship class. Besides, several functions were removed from the Ship class. The Board and GameState classes were also refactored for better organization and readability. This is in line with the objective to improve the game logic and maintainability of the code.