feat/refactor: refactor game state & action validation logic
Refactor the game state and action validation logic across multiple classes. The modifications involve several changes including the addition of detail explanations for exceptions, adding order attribute to action classes, placing movement validation in the GameState class, and updating the method of counting and validating movement points. This is done to increase readability and maintainability in the code, as well as to provide a more intuitive and detailed error logging to improve debugging efficiency. Furthermore, tests are updated to comply with the new changes.